refactor(ast): added id info to ImportFrom (semgrep/semgrep-proprietary#2854)
## What:
This PR makes it so that `ident`s in `ImportFrom`s without aliases also
have an `id_info`.

## Why:


#### Ultimate goal
We would like to be able to analyze the exported symbols from the
`__init__.py`s of a module in Python, which may reference other files.

#### On imports in Python
In Python, `from a import b` is also a declarative statement, which
allows the symbol `b` to be exported from the file. A common pattern is
for `__init__.py` files to import symbols from other files via this
method, such as by having an `__init__.py` which only has the following
contents:
```
from .one import a
from .two import b

# no need to declare a, b, c explicitly!
```

In particular, it would be nice if in the following `__init__.py`:
```
from .one import a
from .two import b
```
we were able to synthesize that the SIG of this file would contain the
symbols `a` and `b`. Currently this looks like:
```
[({ name = { Name.p = ["/Users/brandonspark/playground/test3"]; x = 1707 };
    attrs = []; tparams = []; loc = (Either.Right ()) },
  ModuleDecl)] 
```
which notably only exports the module itself.

Ideally, we could have the `SIG` of this `__init__.py` contain an
`AliasDecl`, which references the proper symbol which is being exported.

#### Constraints and `id_info`
However, to do this, we would need to be able to be holding an `entity`
for the ident, and we cannot do that, because currently, we do not have
`id_info`s associated to our `ImportFrom` idents, when they do not have
an alias.

See the following translations:
```python
from a import b
```
->
```
Pr(
  [DirectiveStmt(
     {d_attrs=[];
      d=ImportFrom((), DottedName([("a", ())]), [(("b", ()), None)]); })])
```

```python
from a import b as c
```
->
```
Pr(
  [DirectiveStmt(
     {d_attrs=[];
      d=ImportFrom((), DottedName([("a", ())]),
          [(("b", ()),
            Some((("c", ()),
                  {id_info_id=2; id_flags=Ref(0);
                   id_resolved_alternative=Ref([]); id_resolved=Ref(None);
                   id_type=Ref(None); id_svalue=Ref(None); })))]);
      })])
```

So in order for us to properly treat `from .one import a` where the `a`
is a proper, resolvable ident in its own right, we have to add an
`id_info` to it.

## How:
Added a new variant, `import_from_kind`, which denotes if the
`ImportFrom` is a direct kind or an aliased kind. Both have exactly one
`id_info` inside of the import.

This PR doesn't percolate the changes to the SAST, as I thought it might
be nice to have these changes in a contained way here, and break things
up into parts.

## Test plan:
We see that `Naming_AST`'s changes seem reasonable:
```
[13:49:43] brandonspark@brandons-mbp λ ~/playground> cat test.py
from a import b as c

x = b
y = c⏎
[13:49:49] brandonspark@brandons-mbp λ ~/playground> sc -dump_named_ast test.py -lang python
!!!This is a proprietary extension of semgrep.!!!
!!!You should not call directly this program.!!!
[00.09][INFO]: Executed as: /Users/brandonspark/deep-semgrep/_build/default/src/main/Main.exe -dump_named_ast test.py -lang python
[00.09][INFO]: Version: 1.102.0
Pr(
  [DirectiveStmt(
     {d_attrs=[];
      d=ImportFrom((), DottedName([("a", ())]),
          [Aliased(("b", ()),
             (("c", ()),
              {id_info_id=2; id_flags=Ref(0);
               id_resolved_alternative=Ref([]);
               id_resolved=Ref(Some((ImportedEntity(["a"; "b"]), 1)));
               id_type=Ref(None); id_svalue=Ref(None); }))]);
      });
   ExprStmt(
     Assign(
       N(
         Id(("x", ()),
           {id_info_id=3; id_flags=Ref(4); id_resolved_alternative=Ref(
            []); id_resolved=Ref(Some((Global, 2))); id_type=Ref(None);
            id_svalue=Ref(None); })), (),
       N(
         Id(("b", ()),
           {id_info_id=4; id_flags=Ref(0); id_resolved_alternative=Ref(
            []); id_resolved=Ref(None); id_type=Ref(None);
            id_svalue=Ref(None); }))), ());
   ExprStmt(
     Assign(
       N(
         Id(("y", ()),
           {id_info_id=5; id_flags=Ref(4); id_resolved_alternative=Ref(
            []); id_resolved=Ref(Some((Global, 3))); id_type=Ref(None);
            id_svalue=Ref(None); })), (),
       N(
         Id(("c", ()),
           {id_info_id=6; id_flags=Ref(0); id_resolved_alternative=Ref(
            []); id_resolved=Ref(Some((ImportedEntity(["a"; "b"]), 1)));
            id_type=Ref(None); id_svalue=Ref(None); }))), ())])
```

Note that both `id_resolved`s are annotated the same way. We also only
name the `c`, but not the `b`, which was not actually imported under
that name.

synced from Pro d7c8e805075c7417d44a722dd09a87527ef89e1c