ci(windows): Enable Windows CI build for semgrep-core (semgrep/semgrep-proprietary#3023)
This PR enables Windows CI builds and contains all the changes required
to get a successful build.

conf-libcurl was locked to an older version to make it work on Windows,
but with the latest opam the latest version works fine. So, we no longer
force use of the older version.

Installing conf-pkg-config reinstalls mingw-w64-shims and the PATH
changes performed by `mingw-w64-shims` need to be available before
installing other packages (conf-libcurl, for instance). So, we install
it before we begin installing other packages.

This commit also makes a bunch of changes to the Windows GitHub action
to get it to work with setup-ocaml@v3. The details of which can be found
in the commit messages.

synced from Pro 9fa620a83dc9838713b062bf57daadd2731d3e00