feat(matching): match parameterized types (semgrep/semgrep-proprietary#3117)
## What:
This PR makes it so we can match parameterized types with unparameterized ones, so long as the names match.

## Why:
This lets us do queries on, for instance, constructions of instances of a class, without having to specify the type arguments to the class as a whole.

## How:
Following the `less-is-ok` philosophy, I just added a case to `Pattern_vs_code`.

## Test plan:
Automated tests.

synced from Pro a2b3879ee46c76eef09683dd0a7a8fb4946cff59